move readManifest to test file
