--- constants/argument_constants.go
@@ -1,9 +1,0 @@
-package constants
-
-const (
-	DEFAULT_CONFIG_FILE_PATH        = "./config.yml"
-	CONFIG_FILE_ARG                 = "config"
-	DEFAULT_LOG_LEVEL               = "DEBUG"
-	LOG_LEVEL_ARG                   = "log"
-	ENABLE_ENV_VAR_HANDLER_FLAG_ARG = "env"
-)
