--- constants/environment_constants.go
@@ -1,5 +1,0 @@
-package constants
-
-const (
-	LOG_LEVEL_ENV_VAR_NAME = "DEPLOYADACTYL_LOGLEVEL"
-)
